requirements.txt
Flask
onnxruntime
Pillow
